fix: Mastodon tootctl を login shell 経由で実行し rbenv を読み込ませる (#44)
cron/periodic の非ログインシェルでは rbenv が初期化されず、
sudo -u mastodon sh -c 'bundle …' が OS 同梱 Ruby にフォールバックする。
Mastodon 4.6 で Ruby が 3系→4系へ上がった結果、bundle が未充足と誤判定され
tootctl（週次/日次/時次の maintenance・media_cleanup・follow）が全て中断していた。

tootctl と bundle check を MastodonTootctl モジュールに集約し、
bash -lc（login shell）経由で実行して mastodon ユーザーの rbenv を読み込ませる。
tools 本体は従来どおり OS Ruby で動く。
